adb everything: refactor the core and functions yet again
now the core no longer prints items. instead, if something fails, a Null value is returned. The functions needed to be updated to reflect this, and the front level examples are updated on demonstrating how to handle the return data

rename the functions in the class. it's an adb class, the functions don't need adb in the name as well